cleanup IDs
